+++ tests/Validator/Number/BetweenValidatorTest.php
@@ -1,8 +1,10 @@
 <?php
+
 declare(strict_types=1);
 
 namespace ExtendsSoftware\ExaPHP\Validator\Number;
 
+use ExtendsSoftware\ExaPHP\Validator\Exception\TemplateNotFound;
 use PHPUnit\Framework\TestCase;
 
 class BetweenValidatorTest extends TestCase
@@ -52,17 +54,18 @@ class BetweenValidatorTest extends TestCase
      *
      * Test that values are valid.
      *
-     * @param float|int $number
+     * @param int $integer
+     *
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::__construct()
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::validate()
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::getTemplates()
      * @dataProvider validInclusiveDataProvider
      */
-    public function testValid(float|int $number): void
+    public function testValid(int $integer): void
     {
         $validator = new BetweenValidator(1, 10);
 
-        $this->assertTrue($validator->validate($number)->isValid());
+        $this->assertTrue($validator->validate($integer)->isValid());
     }
 
     /**
@@ -70,17 +73,19 @@ class BetweenValidatorTest extends TestCase
      *
      * Test that values are invalid.
      *
-     * @param float|int $number
+     * @param int $integer
+     *
+     * @throws TemplateNotFound
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::__construct()
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::validate()
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::getTemplates()
      * @dataProvider invalidInclusiveDataProvider
      */
-    public function testInvalid(float|int $number): void
+    public function testInvalid(int $integer): void
     {
         $validator = new BetweenValidator(1, 10);
 
-        $this->assertFalse($validator->validate($number)->isValid());
+        $this->assertFalse($validator->validate($integer)->isValid());
     }
 
     /**
@@ -88,17 +93,18 @@ class BetweenValidatorTest extends TestCase
      *
      * Test that bound values are invalid when inclusive is false.
      *
-     * @param float|int $number
+     * @param int $integer
+     *
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::__construct()
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::validate()
      * @covers       \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::getTemplates()
      * @dataProvider invalidExclusiveDataProvider
      */
-    public function testExclusive(float|int $number): void
+    public function testExclusive(int $integer): void
     {
         $validator = new BetweenValidator(1, 10, false);
 
-        $this->assertFalse($validator->validate($number)->isValid());
+        $this->assertFalse($validator->validate($integer)->isValid());
     }
 
     /**
@@ -113,6 +119,24 @@ class BetweenValidatorTest extends TestCase
         $validator = new BetweenValidator(2);
         $result = $validator->validate('a');
 
+        $this->assertFalse($result->isValid());
+    }
+
+    /**
+     * Allow string.
+     *
+     * Test that integer value as a string is allowed and will validate.
+     *
+     * @covers \ExtendsSoftware\ExaPHP\Validator\Number\BetweenValidator::validate()
+     */
+    public function testAllowString(): void
+    {
+        $validator = new BetweenValidator(1, 10, allowString: true);
+
+        $result = $validator->validate('5');
+        $this->assertTrue($result->isValid());
+
+        $result = $validator->validate('5.5');
         $this->assertFalse($result->isValid());
     }
 }