+++ CHANGELOG.md
@@ -37,6 +37,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Added RangeValidator to validate if right value is greater than (or equals, when inclusive) left value.
 - Added HTTP renderer support for \Generator response body.
 - Added multibyte flag to LengthValidator to switch between strlen() and mb_strlen().
+- Added allowString flag to BetweenValidator to allow string representations of integers (e.g. in case of a URL value).
 
 ### Changed